# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/Surgeryappointments": {
            "post": {
                "description": "Create Surgeryappointment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create Surgeryappointment",
                "operationId": "create-Surgeryappointment",
                "parameters": [
                    {
                        "description": "Surgeryappointment entity",
                        "name": "Surgeryappointment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Surgeryappointment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Surgeryappointment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
        "/surgeryappointments": {
            "get": {
                "description": "list surgeryappointment entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List surgeryappointment entities",
                "operationId": "list-surgeryappointment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Surgeryappointment"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/surgeryappointments/{id}": {
            "get": {
                "description": "get surgeryappointment by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a surgeryappointment entity by ID",
                "operationId": "get-surgeryappointment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "surgeryappointment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Surgeryappointment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "update surgeryappointment by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a surgeryappointment entity by ID",
                "operationId": "update-surgeryappointment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Surgeryappointment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "surgeryappointment entity",
                        "name": "surgeryappointment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Surgeryappointment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Surgeryappointment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get surgeryappointment by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a surgeryappointment entity by ID",
                "operationId": "delete-surgeryappointment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Surgeryappointment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.Surgeryappointment": {
            "type": "object",
            "properties": {
                "appointtime": {
                    "type": "string"
                },
                "patientid": {
                    "type": "integer"
                },
                "personelid": {
                    "type": "integer"
                },
                "surgerytypeid": {
                    "type": "integer"
                }
            }
        },
        "ent.Antenatalinformation": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the AntenatalinformationQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.AntenatalinformationEdges"
                },
                "gestationalage": {
                    "description": "Gestationalage holds the value of the \"gestationalage\" field.",
                    "type": "integer"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "time": {
                    "description": "Time holds the value of the \"time\" field.",
                    "type": "string"
                }
            }
        },
        "ent.AntenatalinformationEdges": {
            "type": "object",
            "properties": {
                "patient": {
                    "description": "Patient holds the value of the Patient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patient"
                },
                "personnel": {
                    "description": "Personnel holds the value of the Personnel edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Personnel"
                },
                "pregnancystatus": {
                    "description": "Pregnancystatus holds the value of the Pregnancystatus edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Pregnancystatus"
                },
                "risks": {
                    "description": "Risks holds the value of the Risks edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Risks"
                }
            }
        },
        "ent.Bonedisease": {
            "type": "object",
            "properties": {
                "addedTime": {
                    "description": "AddedTime holds the value of the \"addedTime\" field.",
                    "type": "string"
                },
                "advice": {
                    "description": "Advice holds the value of the \"advice\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the BonediseaseQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.BonediseaseEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.BonediseaseEdges": {
            "type": "object",
            "properties": {
                "patient": {
                    "description": "Patient holds the value of the patient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patient"
                },
                "personnel": {
                    "description": "Personnel holds the value of the personnel edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Personnel"
                },
                "remedy": {
                    "description": "Remedy holds the value of the remedy edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Remedy"
                }
            }
        },
        "ent.Checksymptoms": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "Date holds the value of the \"date\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ChecksymptomsQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ChecksymptomsEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.ChecksymptomsEdges": {
            "type": "object",
            "properties": {
                "disease": {
                    "description": "Disease holds the value of the disease edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Disease"
                },
                "doctorordersheet": {
                    "description": "Doctorordersheet holds the value of the doctorordersheet edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DoctorOrderSheet"
                },
                "patient": {
                    "description": "Patient holds the value of the patient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patient"
                },
                "personnel": {
                    "description": "Personnel holds the value of the personnel edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Personnel"
                }
            }
        },
        "ent.Dentalappointment": {
            "type": "object",
            "properties": {
                "appointtime": {
                    "description": "Appointtime holds the value of the \"appointtime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DentalappointmentQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DentalappointmentEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.DentalappointmentEdges": {
            "type": "object",
            "properties": {
                "dentalkind": {
                    "description": "Dentalkind holds the value of the Dentalkind edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Dentalkind"
                },
                "patient": {
                    "description": "Patient holds the value of the Patient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patient"
                },
                "personnel": {
                    "description": "Personnel holds the value of the Personnel edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Personnel"
                }
            }
        },
...
        "ent.Patient": {
            "type": "object",
            "properties": {
                "birthday": {
                    "description": "Birthday holds the value of the \"birthday\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PatientQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PatientEdges"
                },
                "gender": {
                    "description": "Gender holds the value of the \"gender\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                }
            }
        },
        "ent.PatientEdges": {
            "type": "object",
            "properties": {
                "antenatalinformation": {
                    "description": "Antenatalinformation holds the value of the Antenatalinformation edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Antenatalinformation"
                    }
                },
                "bonedisease": {
                    "description": "Bonedisease holds the value of the Bonedisease edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Bonedisease"
                    }
                },
                "checksymptoms": {
                    "description": "Checksymptoms holds the value of the Checksymptoms edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Checksymptoms"
                    }
                },
                "dentalappointment": {
                    "description": "Dentalappointment holds the value of the Dentalappointment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Dentalappointment"
                    }
                },
                "physicaltherapyrecord": {
                    "description": "Physicaltherapyrecord holds the value of the physicaltherapyrecord edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Physicaltherapyrecord"
                    }
                },
                "surgeryappointment": {
                    "description": "Surgeryappointment holds the value of the Surgeryappointment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Surgeryappointment"
                    }
                }
            }
        },
        "ent.Personnel": {
            "type": "object",
            "properties": {
                "department": {
                    "description": "Department holds the value of the \"department\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PersonnelQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PersonnelEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "password": {
                    "description": "Password holds the value of the \"password\" field.",
                    "type": "string"
                },
                "user": {
                    "description": "User holds the value of the \"user\" field.",
                    "type": "string"
                }
            }
        },
        "ent.PersonnelEdges": {
            "type": "object",
            "properties": {
                "antenatalinformation": {
                    "description": "Antenatalinformation holds the value of the Antenatalinformation edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Antenatalinformation"
                    }
                },
                "bonedisease": {
                    "description": "Bonedisease holds the value of the Bonedisease edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Bonedisease"
                    }
                },
                "checksymptoms": {
                    "description": "Checksymptoms holds the value of the Checksymptoms edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Checksymptoms"
                    }
                },
                "dentalappointment": {
                    "description": "Dentalappointment holds the value of the Dentalappointment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Dentalappointment"
                    }
                },
                "physicaltherapyrecord": {
                    "description": "Physicaltherapyrecord holds the value of the physicaltherapyrecord edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Physicaltherapyrecord"
                    }
                },
                "surgeryappointment": {
                    "description": "Surgeryappointment holds the value of the Surgeryappointment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Surgeryappointment"
                    }
                }
            }
        },
        "ent.Physicaltherapyrecord": {
            "type": "object",
            "properties": {
                "added_time": {
                    "description": "AddedTime holds the value of the \"added_time\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PhysicaltherapyrecordQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PhysicaltherapyrecordEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Surgeryappointment": {
            "type": "object",
            "properties": {
                "appoint_time": {
                    "description": "AppointTime holds the value of the \"appoint_time\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the SurgeryappointmentQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.SurgeryappointmentEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.SurgeryappointmentEdges": {
            "type": "object",
            "properties": {
                "patient": {
                    "description": "Patient holds the value of the Patient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patient"
                },
                "personnel": {
                    "description": "Personnel holds the value of the Personnel edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Personnel"
                },
                "surgerytype": {
                    "description": "Surgerytype holds the value of the Surgerytype edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Surgerytype"
                }
            }
        },
        "ent.Surgerytype": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the SurgerytypeQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.SurgerytypeEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "typename": {
                    "description": "Typename holds the value of the \"typename\" field.",
                    "type": "string"
                }
            }
        },
        "ent.SurgerytypeEdges": {
            "type": "object",
            "properties": {
                "surgeryappointment": {
                    "description": "Surgeryappointment holds the value of the Surgeryappointment edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Surgeryappointment"
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "authorizationUrl": "",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "authorizationUrl": "",
...
}